% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler.R
\name{Bmggm}
\alias{Bmggm}
\title{Bayesian multiple Gaussian graphical models by MCMC.}
\usage{
Bmggm(dat, options, PriorPar, InitVal)
}
\arguments{
\item{dat}{a list of objets:
n: number of observations.
p: dimension of each pathway.
K: number of pathways.
z_P: indicator vector of genes membership
P: dimension of the data.}

\item{options}{a list of objets:
burnin: number of MCMC iterations before burnin.
nmc: number of MCMC iterations after burnin.}

\item{PriorPar}{a list of objets:
a: shape1 parameter for Theta for off-digonal block.
b: shape2 parameter for Theta for off-digonal block.
a0: shape1 parameter for Theta for digonal block.
b0: shape2 parameter for Theta for digonal block.
eps: rate parameter for v0^2.
delta: shape parameter for v0^2.
c: the parameter for decision boundary of spike-and-slab.
Theta: a K x K initial graph PPI matrix.}

\item{InitVal}{a list of objets:
mu: intercept term.
sigma2: overall noise level, same across groups.
Beta: a P x P initial coefficient matrix. 
adj: a P x P initial adjacency matrix.}
}
\value{
a list of objets:     
 Beta_save: p x p x K x nmc sample of coefficient matrix
 adj_save: p x p x K x nmc sample of adjacency matrix
 Theta_save: K x K x nmc sample of graph similarity matrix
}
\description{
Bayesian multiple Gaussian graphical models by MCMC.
}
