% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FixMatrix}
\alias{FixMatrix}
\title{Fix to ensure positive definiteness from Danaher et al.
Divide each off-diagonal element by sum of absolute values of off-diagonal elements in its row.}
\usage{
FixMatrix(A, denom_factor)
}
\arguments{
\item{A}{a raw matrix.}

\item{denom_factor}{scale parameter.}
}
\value{
B a symmetric and positive definite matrix.
}
\description{
Fix to ensure positive definiteness from Danaher et al.
Divide each off-diagonal element by sum of absolute values of off-diagonal elements in its row.
}
